Show visits over past week/month/year
Update the stats page to show the number of visits over the past
week/month/year as well as the averages for those periods.
The admin page now retrieves the values for the averages above.
Changed the admin index page to evaluate values at haml compilation
time rather than relying on JS to fill them in.