Complete the tests for editing teams
Some of these tests work when they want to, will have to look into
them further.
Realised that some of the names generated by Faker had apostophes
(') in them, which was prematurely closing string, causing erros
when testing (and in general).
For now I've replaced the single quoted strings with double quoted
strings, but I will later switch back to single quotes
and implement something to add the backslash
(\) character in front of every apostrophe '.